Implement satcheljs-promise (#46)
This introduces satcheljs-promise, which is a middleware that overrides the default Promise's `then` and `catch` methods to create actions out of the fulfilled/rejected callbacks.  This makes it simple to write actions that do asynchronous work, either via promises or async/await, without having to manually wrap everything that happens asynchronously in separate actions.