Change the way we expose action type (#50)
The way we expose an actionType property off actions is problematic because in order to generate definition files you need to export the `Action` type anywhere you export an action.  This changes the way we do it: now we stamp a private property on the action (i.e. don't expose it via an interface) and have Satchel export a getter function that can read and return that private property.