Update dependencies (#60)
With newer versions of some of our typings dependencies our build was broken.  I had to update TypeScript (to support default type parameters in generics) and some of the typings, so I ended up refreshing a bunch of our dependencies.  I also switched our typings to ~ dependencies to hopefully avoid this sort of break in the future.