Rename several APIs (#78)
This renames the following APIs (and updates the README):

`simpleMutator` => `mutatorAction`
`simpleOrchestrator` => `orchestratorAction`
`boundActionCreator` => `action`

For the "simple" APIs, the intent is to make the name a little more specific about what the API is actually creating.  We're renaming `boundActionCreator` because we want the simplest, default kind of action creator to be the one that automatically dispatches.