Pipe promises returned from orchestrators through to middleware (#79)
This is necessary because orchestrators can be async and sometimes middleware wants to know when the orchestrator has actually completed.  Since there can be multiple orchestrators subscribed to an action, we'll aggregate the returned promises into one for the purposes of middleware.  Besides that little bit of logic, the bulk of the change here is a bunch of type updates.