Update to MobX v4 (#98)
This updates Satchel to the latest versions of `mobx` and `mobx-react`.  There were a number of breaking API changes, including some (as far as I can tell) undocumented ones in the spying infrastructure.  This will be in v4 of Satchel since consumers will be forced to take a new major version of MobX.