Wrap dispatch in a transaction (#106)
For various historical reasons dispatch hasn't been wrapped in `transaction` since an early version of Satchel.  This means that every subscriber to an action executes in its own transaction, possibly causing multiple unnecessary renders.  Now that we've updated to MobX v4, it's possible to use `transaction` again.

The change to actual source code is trivial; the rest of the change is test code to validate that we're handling the transaction correctly.