Fix global reference (#151)
`global` is a value in NodeJS, and is polyfilled by webpack, but it is not defined in any browser context. `globalThis` is meant to act as the global reference in all contexts, but is only implemented in Edge/Chrome >= 71, Firefox >= 65, Safari >= 12.1, and NodeJS >= 12

This change replaces all references to `global` with `globalObject`, a value that is initialized to point to either `globalThis`, `window`, or `global`, depending on which is defined.